requirements.txt
colorlog
homeassistant
ruff
pytest-asyncio
